requirements.txt
streamlit
transformers
sentence-transformers
torch
faiss-cpu
scikit-learn
numpy
pandas
tqdm
